+++ src/org/molgenis/generators/csv/CsvExportGen.java
@@ -2,8 +2,10 @@ package org.molgenis.generators.csv;
 
 import java.io.File;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
+import java.nio.charset.Charset;
 import java.util.List;
 import java.util.Map;
 
@@ -19,35 +21,61 @@ import freemarker.template.Template;
 public class CsvExportGen extends MySqlCreateClassPerTableGen
 {
 	public static transient final Logger logger = Logger.getLogger(CsvExportGen.class);
-	
+
 	@Override
 	public String getDescription()
 	{
 		return "Generates CsvExport";
 	}
-	
+
 	@Override
-	public void generate(Model model, MolgenisOptions options)
-			throws Exception
+	public void generate(Model model, MolgenisOptions options) throws Exception
 	{
-		Template template = createTemplate( "/"+this.getClass().getSimpleName()+".java.ftl" );
+		Template template = createTemplate("/" + this.getClass().getSimpleName() + ".java.ftl");
 		Map<String, Object> templateArgs = createTemplateArguments(options);
-		
+
 		List<Entity> entityList = model.getEntities();
-		entityList = MolgenisModel.sortEntitiesByDependency(entityList,model); //side effect?
-		//String packageName = this.getClass().getPackage().toString().substring(Generator.class.getPackage().toString().length());
+		entityList = MolgenisModel.sortEntitiesByDependency(entityList, model); // side
+																				// effect?
+		File target = new File(this.getSourcePath(options) + APP_DIR + "/CsvExport.java");
+		boolean created = target.getParentFile().mkdirs();
+		if (!created && !target.getParentFile().exists())
+		{
+			throw new IOException("could not create " + target.getParentFile());
+		}
 
-
-		File target = new File( this.getSourcePath(options) + APP_DIR+"/CsvExport.java" );
-		target.getParentFile().mkdirs();
-		
-		templateArgs.put( "model", model );
-		templateArgs.put("entities",entityList);
+		templateArgs.put("model", model);
+		templateArgs.put("entities", entityList);
 		templateArgs.put("package", APP_DIR);
-		OutputStream targetOut = new FileOutputStream( target );
-		template.process( templateArgs, new OutputStreamWriter( targetOut ) );
+		OutputStream targetOut = new FileOutputStream(target);
+		template.process(templateArgs, new OutputStreamWriter(targetOut, Charset.forName("UTF-8")));
 		targetOut.close();
-		
+
+		logger.info("generated " + target);
+	}
+
+	public void generate(Model model, MolgenisOptions options, OutputStream os) throws Exception
+	{
+		Template template = createTemplate("/" + this.getClass().getSimpleName() + ".java.ftl");
+		Map<String, Object> templateArgs = createTemplateArguments(options);
+
+		List<Entity> entityList = model.getEntities();
+		entityList = MolgenisModel.sortEntitiesByDependency(entityList, model); // side
+																				// effect?
+		File target = new File("/git/molgenis/generated/java/org/molgenis/generators/csv/CsvExport.java");
+		boolean created = target.getParentFile().mkdirs();
+		if (!created && !target.getParentFile().exists())
+		{
+			throw new IOException("could not create " + target.getParentFile());
+		}
+
+		templateArgs.put("model", model);
+		templateArgs.put("entities", entityList);
+		templateArgs.put("package", "org.molgenis.generators.csv");
+		OutputStream targetOut = new FileOutputStream(target);
+		template.process(templateArgs, new OutputStreamWriter(targetOut, Charset.forName("UTF-8")));
+		targetOut.close();
+
 		logger.info("generated " + target);
 	}
 }
