+++ src/org/molgenis/generators/db/MapperDecoratorGen.java
@@ -2,8 +2,10 @@ package org.molgenis.generators.db;
 
 import java.io.File;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
+import java.nio.charset.Charset;
 import java.util.Map;
 
 import org.molgenis.MolgenisOptions;
@@ -42,38 +44,44 @@ public class MapperDecoratorGen extends ForEachEntityGenerator
 				if (entity.getDecorator() != null)
 				{
 					String fullKlazzName = entity.getDecorator();
-					
+
 					String packageName = fullKlazzName;
-					if(fullKlazzName.contains("."))
-						packageName = fullKlazzName.substring(0, fullKlazzName.lastIndexOf("."));
-					
+					if (fullKlazzName.contains(".")) packageName = fullKlazzName.substring(0,
+							fullKlazzName.lastIndexOf("."));
+
 					String shortKlazzName = fullKlazzName;
-					if(fullKlazzName.contains("."))
-						shortKlazzName = fullKlazzName.substring(fullKlazzName.lastIndexOf(".") + 1);
+					if (fullKlazzName.contains(".")) shortKlazzName = fullKlazzName.substring(fullKlazzName
+							.lastIndexOf(".") + 1);
 
 					File targetDir = new File(this.getHandWrittenPath(options) + "/" + packageName.replace(".", "/"));
-					targetDir.mkdirs();
+					boolean created = targetDir.mkdirs();
+					if (!created && !targetDir.exists())
+					{
+						throw new IOException("could not create " + targetDir);
+					}
 
 					File targetFile = new File(this.getHandWrittenPath(options) + "/" + fullKlazzName.replace(".", "/")
 							+ ".java");
 					// only generate if the file doesn't exist
 					if (!targetFile.exists())
 					{
-						templateArgs.put("entityClass", entity.getNamespace()+"." + 
-								GeneratorHelper.getJavaName(entity.getName()));
+						templateArgs.put("entityClass",
+								entity.getNamespace() + "." + GeneratorHelper.getJavaName(entity.getName()));
 						templateArgs.put("clazzName", shortKlazzName);
 						templateArgs.put("entity", entity);
 						templateArgs.put("model", model);
 						// templateArgs.put("db_driver", options.db_driver);
 						templateArgs.put("template", template.getName());
-						templateArgs.put("file", packageName.replace(".", "/") + "/"
-								+ GeneratorHelper.getJavaName(entity.getName()) + getType() + getExtension());
+						templateArgs.put("file",
+								packageName.replace(".", "/") + "/" + GeneratorHelper.getJavaName(entity.getName())
+										+ getType() + getExtension());
 						templateArgs.put("package", packageName);
-						templateArgs.put("databaseImp", options.mapper_implementation.equals(MolgenisOptions.MapperImplementation.JPA) ? "jpa" : "jdbc");
+						templateArgs.put("databaseImp", options.mapper_implementation
+								.equals(MolgenisOptions.MapperImplementation.JPA) ? "jpa" : "jdbc");
 
 						OutputStream targetOut = new FileOutputStream(targetFile);
 
-						template.process(templateArgs, new OutputStreamWriter(targetOut));
+						template.process(templateArgs, new OutputStreamWriter(targetOut, Charset.forName("UTF-8")));
 						targetOut.close();
 
 						// logger.info("generated " +
