+++ src/org/molgenis/generators/doc/TableDocGen.java
@@ -2,8 +2,10 @@ package org.molgenis.generators.doc;
 
 import java.io.File;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
+import java.nio.charset.Charset;
 import java.util.Map;
 
 import org.apache.log4j.Logger;
@@ -22,22 +24,25 @@ public class TableDocGen extends Generator
 	{
 		return "Generates one documentation file describing all generated tables.";
 	}
-	
+
 	@Override
-	public void generate(Model model, MolgenisOptions options)
-			throws Exception
+	public void generate(Model model, MolgenisOptions options) throws Exception
 	{
-		Template template = createTemplate( "/"+getClass().getSimpleName()+".java.ftl" );
+		Template template = createTemplate("/" + getClass().getSimpleName() + ".java.ftl");
 		Map<String, Object> templateArgs = createTemplateArguments(options);
 
-		File target = new File(this.getDocumentationPath( options ) +"/tabledoc.html");		
-		target.getParentFile().mkdirs();
-		
-		templateArgs.put("model", model );
-		OutputStream targetOut = new FileOutputStream( target );
-		template.process( templateArgs, new OutputStreamWriter( targetOut ) );
+		File target = new File(this.getDocumentationPath(options) + "/tabledoc.html");
+		boolean created = target.getParentFile().mkdirs();
+		if (!created && !target.getParentFile().exists())
+		{
+			throw new IOException("could not create " + target.getParentFile());
+		}
+
+		templateArgs.put("model", model);
+		OutputStream targetOut = new FileOutputStream(target);
+		template.process(templateArgs, new OutputStreamWriter(targetOut, Charset.forName("UTF-8")));
 		targetOut.close();
-		
+
 		logger.info("generated " + target);
 	}
