+++ src/org/molgenis/generators/excel/ImportWizardExcelPrognosisGen.java
@@ -2,8 +2,10 @@ package org.molgenis.generators.excel;
 
 import java.io.File;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
+import java.nio.charset.Charset;
 import java.util.List;
 import java.util.Map;
 
@@ -19,34 +21,37 @@ import freemarker.template.Template;
 public class ImportWizardExcelPrognosisGen extends MySqlCreateClassPerTableGen
 {
 	public static final transient Logger logger = Logger.getLogger(ImportWizardExcelPrognosisGen.class);
-	
+
 	@Override
 	public String getDescription()
 	{
 		return "Generates a prognosis report maker for the generic data import wizard using Excel files";
 	}
-	
-	@Override
-	public void generate(Model model, MolgenisOptions options)
-			throws Exception
-	{
-		Template template = createTemplate( "/"+this.getClass().getSimpleName()+".java.ftl" );
-		Map<String, Object> templateArgs = createTemplateArguments(options);
-		
-		List<Entity> entityList = model.getEntities();
-		entityList = MolgenisModel.sortEntitiesByDependency(entityList,model); //side effect?
-		//String packageName = this.getClass().getPackage().toString().substring(Generator.class.getPackage().toString().length());
 
-		File target = new File( this.getSourcePath(options) + APP_DIR+"/ImportWizardExcelPrognosis.java" );
-		target.getParentFile().mkdirs();
-		
-		templateArgs.put( "model", model );
-		templateArgs.put("entities",entityList);
+	@Override
+	public void generate(Model model, MolgenisOptions options) throws Exception
+	{
+		Template template = createTemplate("/" + this.getClass().getSimpleName() + ".java.ftl");
+		Map<String, Object> templateArgs = createTemplateArguments(options);
+
+		List<Entity> entityList = model.getEntities();
+		entityList = MolgenisModel.sortEntitiesByDependency(entityList, model); // side
+																				// effect?
+
+		File target = new File(this.getSourcePath(options) + APP_DIR + "/ImportWizardExcelPrognosis.java");
+		boolean created = target.getParentFile().mkdirs();
+		if (!created && !target.getParentFile().exists())
+		{
+			throw new IOException("could not create " + target.getParentFile());
+		}
+
+		templateArgs.put("model", model);
+		templateArgs.put("entities", entityList);
 		templateArgs.put("package", APP_DIR);
-		OutputStream targetOut = new FileOutputStream( target );
-		template.process( templateArgs, new OutputStreamWriter( targetOut ) );
+		OutputStream targetOut = new FileOutputStream(target);
+		template.process(templateArgs, new OutputStreamWriter(targetOut, Charset.forName("UTF-8")));
 		targetOut.close();
-		
+
 		logger.info("generated " + target);
 	}
 }
