+++ src/org/molgenis/generators/server/SoapApiGen.java
@@ -2,8 +2,10 @@ package org.molgenis.generators.server;
 
 import java.io.File;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
+import java.nio.charset.Charset;
 import java.util.List;
 import java.util.Map;
 
@@ -26,30 +28,32 @@ public class SoapApiGen extends Generator
 	{
 		return "Generates soap service interfaces for each entity.";
 	}
-	
+
 	@Override
-	public void generate(Model model, MolgenisOptions options)
-			throws Exception
+	public void generate(Model model, MolgenisOptions options) throws Exception
 	{
-		Template template = createTemplate( "/"+this.getClass().getSimpleName()+".java.ftl" );
+		Template template = createTemplate("/" + this.getClass().getSimpleName() + ".java.ftl");
 		Map<String, Object> templateArgs = createTemplateArguments(options);
-		
+
 		List<Entity> entityList = model.getEntities();
 		List<Method> methodList = model.getMethods();
 
+		File target = new File(this.getSourcePath(options) + APP_DIR + "/servlet/SoapApi.java");
+		boolean created = target.getParentFile().mkdirs();
+		if (!created && !target.getParentFile().exists())
+		{
+			throw new IOException("could not create " + target.getParentFile());
+		}
 
-		File target = new File( this.getSourcePath(options) + APP_DIR + "/servlet/SoapApi.java" );
-		target.getParentFile().mkdirs();
-		
-		templateArgs.put("model", model );
+		templateArgs.put("model", model);
 		templateArgs.put("methods", methodList);
 		templateArgs.put("entities", entityList);
 		templateArgs.put("helper", new GeneratorHelper(null));
-		templateArgs.put("package", APP_DIR+".servlet");
-		OutputStream targetOut = new FileOutputStream( target );
-		template.process( templateArgs, new OutputStreamWriter( targetOut ) );
+		templateArgs.put("package", APP_DIR + ".servlet");
+		OutputStream targetOut = new FileOutputStream(target);
+		template.process(templateArgs, new OutputStreamWriter(targetOut, Charset.forName("UTF-8")));
 		targetOut.close();
-		
+
 		logger.info("generated " + target);
 	}
 }
