+++ src/org/molgenis/generators/sql/HSqlCreateSubclassPerTableGen.java
@@ -3,8 +3,10 @@ package org.molgenis.generators.sql;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
+import java.nio.charset.Charset;
 import java.sql.Connection;
 import java.sql.Statement;
 import java.util.List;
@@ -25,7 +27,7 @@ public class HSqlCreateSubclassPerTableGen extends Generator
 	public static final transient Logger logger = Logger.getLogger(HSqlCreateSubclassPerTableGen.class);
 
 	@Override
-	public void generate( Model model, MolgenisOptions options ) throws Exception
+	public void generate(Model model, MolgenisOptions options) throws Exception
 	{
 		// create an hsqldb connection
 		BasicDataSource data_src = new BasicDataSource();
@@ -38,19 +40,24 @@ public class HSqlCreateSubclassPerTableGen extends Generator
 
 		Connection conn = data_src.getConnection();
 		Statement stmt = null;
-		
-		//remove existing database
-		
+
+		// remove existing database
+
 		// create generator
-		Template template = this.createTemplate(this.getClass().getSimpleName()+".hsql.ftl");
+		Template template = this.createTemplate(this.getClass().getSimpleName() + ".hsql.ftl");
 		Map<String, Object> templateArgs = createTemplateArguments(options);
-		
-		//Output file for debug
-		File target = new File( this.getSqlPath(options) + "/create_tables.sql" );
-		target.getParentFile().mkdirs();
-		OutputStream targetOut = new FileOutputStream( target );
+
+		// Output file for debug
+		File target = new File(this.getSqlPath(options) + "/create_tables.sql");
+		boolean created = target.getParentFile().mkdirs();
+		if (!created && !target.getParentFile().exists())
+		{
+			throw new IOException("could not create " + target.getParentFile());
+		}
+
+		OutputStream targetOut = new FileOutputStream(target);
 		List<Entity> sortedlist = model.getEntities();
-		sortedlist = MolgenisModel.sortEntitiesByDependency(sortedlist,model);
+		sortedlist = MolgenisModel.sortEntitiesByDependency(sortedlist, model);
 
 		// create arguments
 		templateArgs.put("entities", sortedlist);
@@ -58,21 +65,25 @@ public class HSqlCreateSubclassPerTableGen extends Generator
 
 		// generate
 		ByteArrayOutputStream out = new ByteArrayOutputStream();
-		try{
-			template.process(templateArgs, new OutputStreamWriter(out));
-			//Write to file for debug
+		try
+		{
+			template.process(templateArgs, new OutputStreamWriter(out, Charset.forName("UTF-8")));
+			// Write to file for debug
 			targetOut.write(out.toByteArray());
-			
-			//Update the Hsql database
+
+			// Update the Hsql database
 			stmt = conn.createStatement();
-			stmt.executeUpdate(out.toString());
-		}catch(Exception e){
-			logger.debug("Something wrong with Code:" + out.toString() +" \n Error:" + e.getMessage());
+			stmt.executeUpdate(out.toString("UTF-8"));
+		}
+		catch (Exception e)
+		{
+			logger.debug("Something wrong with Code:" + out.toString() + " \n Error:" + e.getMessage());
 			e.printStackTrace();
 			System.exit(-1);
 		}
-		
-		for(Entity e : sortedlist){
+
+		for (Entity e : sortedlist)
+		{
 			logger.debug("Created hsql table for : " + e.getName());
 		}
 		stmt.executeUpdate("SHUTDOWN");
