+++ src/org/molgenis/generators/sql/OracleCreateSubclassPerTableGen.java
@@ -2,8 +2,10 @@ package org.molgenis.generators.sql;
 
 import java.io.File;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
+import java.nio.charset.Charset;
 import java.util.List;
 import java.util.Map;
 
@@ -25,27 +27,30 @@ public class OracleCreateSubclassPerTableGen extends Generator
 	{
 		return "Generates create tables and views for each entity.";
 	}
-	
+
 	@Override
-	public void generate(Model model, MolgenisOptions options)
-			throws Exception
+	public void generate(Model model, MolgenisOptions options) throws Exception
 	{
-		Template template = createTemplate( "/"+this.getClass().getSimpleName()+".sql.ftl" );
+		Template template = createTemplate("/" + this.getClass().getSimpleName() + ".sql.ftl");
 		Map<String, Object> templateArgs = createTemplateArguments(options);
-		
+
 		List<Entity> entityList = model.getEntities();
-		entityList = MolgenisModel.sortEntitiesByDependency(entityList,model); //side effect?
+		entityList = MolgenisModel.sortEntitiesByDependency(entityList, model); // side
+																				// effect?
 
+		File target = new File(this.getSqlPath(options) + "/create_tables.sql");
+		boolean created = target.getParentFile().mkdirs();
+		if (!created && !target.getParentFile().exists())
+		{
+			throw new IOException("could not create " + target.getParentFile());
+		}
 
-		File target = new File( this.getSqlPath(options) + "/create_tables.sql" );
-		target.getParentFile().mkdirs();
-		
-		templateArgs.put( "model", model );
-		templateArgs.put("entities",entityList);
-		OutputStream targetOut = new FileOutputStream( target );
-		template.process( templateArgs, new OutputStreamWriter( targetOut ) );
+		templateArgs.put("model", model);
+		templateArgs.put("entities", entityList);
+		OutputStream targetOut = new FileOutputStream(target);
+		template.process(templateArgs, new OutputStreamWriter(targetOut, Charset.forName("UTF-8")));
 		targetOut.close();
-		
+
 		logger.info("generated " + target);
 	}
 }
